reserve/api/restaurants.py
from rest_framework import serializers, viewsets
from reserve.models import Restaurant


class RestaurantSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Restaurant
        fields = ['url', 'name', 'address', 'rating', 'description', 'image', 'owned_by']


class RestaurantViewSet(viewsets.ModelViewSet):
    queryset = Restaurant.objects.all()
    serializer_class = RestaurantSerializer
