requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
Pillow==9.1.1
django-bootstrap4==22.1
django-static-fontawesome==5.15.4.0
django-betterforms==1.2
djangorestframework==3.13.1
